backend/nexus/pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "ProjectX backend — AI video interview platform"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",

    # Data validation & settings
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",

    # Database
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",

    # Task queue
    "dramatiq[redis]>=1.17,<2",

    # Redis (session state, cache)
    "redis>=5.2,<6",

    # Auth
    "PyJWT[crypto]>=2.10,<3",

    # HTTP client (ATS adapters, external APIs)
    "httpx>=0.28,<1",

    # Logging & observability
    "structlog>=24.4,<25",
    "sentry-sdk[fastapi]>=2.19,<3",
    "opentelemetry-api>=1.29,<2",
    "opentelemetry-sdk>=1.29,<2",
    "opentelemetry-instrumentation-fastapi>=0.50b,<1",

    # LLM observability (self-hosted)
    "langfuse>=2.56,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.25,<1",
    "pytest-cov>=6.0,<7",
    "httpx",           # TestClient transport
    "ruff>=0.8,<1",
    "mypy>=1.14,<2",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "ASYNC"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
